chore: Attempt to fix tests by starting single servers before
