+++ modules/redis-client/src/test/scala/io/renku/redis/client/util/RedisSpec.scala
@@ -28,6 +28,7 @@ import dev.profunktor.redis4cats.{Redis, RedisCommands}
 import io.lettuce.core.RedisConnectionException
 import io.renku.queue.client.QueueClient
 import io.renku.redis.client.RedisQueueClient
+import io.renku.servers.RedisServer
 
 trait RedisSpec:
   self: munit.Suite =>
